requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
asyncio
aiohttp
azure.eventhub
azure.identity
python-dotenv
confluent-kafka
azure-identity